everything's a double now. other tiny fixes.

+++ trunk/GA_TestBench/src/ga_testbench/Solver.java
@@ -6,10 +6,10 @@ interface Solver<T extends Individual> {
 
     public void setMaxGenerations(int max);
 
-    public void setTimeLimit(float seconds);
+    public void setTimeLimit(double seconds);
 
-    public void setNextGenerationProportions(float copies, float mutations,
-            float crossovers, float randoms);
+    public void setNextGenerationProportions(double copies, double mutations,
+            double crossovers, double randoms);
 
     /* This might provide some statistics to the screen at every nth generation
      * or something like that
